app/requirements.txt
Django==4.1.7
django-allauth==0.53.1
django-crispy-forms==2.0
django-request==1.6.2
gunicorn==20.1.0
crispy-bootstrap5==0.7
# xhtml2pdf==0.2.9
django-filter==23.1
# reportlab==3.6.12
psycopg2-binary==2.9.5
